chore: refactor WagmiWalletCheckoutTest
